nlp/sentiment-engine/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-sentiment-engine"
version = "1.0.0"
description = "Comprehensive Sentiment Aggregation Engine for Crypto Trading with "
authors = [
 {name = "ML Team", email = "ml@example.com"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
 "transformers>=4.35.0",
 "torch>=2.1.0",
 "fastapi>=0.104.0",
 "uvicorn[standard]>=0.24.0",
 "pydantic>=2.5.0",
 "asyncpg>=0.29.0",
 "aioredis>=2.0.1",
 "kafka-python>=2.0.2",
 "websockets>=12.0",
 "nltk>=3.8.1",
 "spacy>=3.7.0",
 "textblob>=0.17.1",
 "vaderSentiment>=3.3.2",
 "tweepy>=4.14.0",
 "praw>=7.7.1",
 "feedparser>=6.0.10",
 "telethon>=1.30.3",
 "discord.py>=2.3.2",
 "scikit-learn>=1.3.0",
 "pandas>=2.1.0",
 "numpy>=1.25.0",
 "scipy>=1.11.0",
 "matplotlib>=3.8.0",
 "seaborn>=0.13.0",
 "plotly>=5.17.0",
 "requests>=2.31.0",
 "aiohttp>=3.9.0",
 "beautifulsoup4>=4.12.2",
 "python-dateutil>=2.8.2",
 "pytz>=2023.3",
 "structlog>=23.2.0",
 "prometheus-client>=0.19.0",
 "opentelemetry-api>=1.21.0",
 "opentelemetry-sdk>=1.21.0",
 "faiss-cpu>=1.7.4",
 "sentence-transformers>=2.2.2",
 "tensorflow>=2.15.0",
]

[project.optional-dependencies]
dev = [
 "pytest>=7.4.0",
 "pytest-asyncio>=0.21.0",
 "pytest-cov>=4.1.0",
 "pytest-mock>=3.12.0",
 "flake8>=6.1.0",
 "mypy>=1.7.0",
 "black>=23.10.0",
 "isort>=5.12.0",
]
gpu = [
 "torch[cuda]>=2.1.0",
 "faiss-gpu>=1.7.4",
]

[project.urls]
Homepage = "https://github.com/vlad/ml-sentiment-engine"
Repository = "https://github.com/vlad/ml-sentiment-engine.git"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
 # directories
 \.eggs
 | \.git
 | \.hg
 | \.mypy_cache
 | \.tox
 | \.venv
 | build
 | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
 "--strict-markers",
 "--strict-config",
 "--verbose",
]
markers = [
 "slow: marks tests as slow (deselect with '-m \"not slow\"')",
 "integration: marks tests as integration tests",
 "unit: marks tests as unit tests",
]